added confirm before post is deleted

+++ public/js/post.js
@@ -19,16 +19,20 @@ const addProject = async (event) => {
     }
   }
 };
+
 const delButtonHandler = async (event) => {
   if (event.target.hasAttribute("data-id")) {
     const id = event.target.getAttribute("data-id");
-    const response = await fetch(`/api/posts/${id}`, {
-      method: "DELETE",
-    });
-    if (response.ok) {
-      document.location.replace("/profile");
-    } else {
-      alert("Failed to delete project");
+    const confirmed = confirm("are you sure you ewant to delete this post?");
+    if (confirmed) {
+      const response = await fetch(`/api/posts/${id}`, {
+        method: "DELETE",
+      });
+      if (response.ok) {
+        document.location.replace("/profile");
+      } else {
+        alert("Failed to delete project");
+      }
     }
   }
 };
